Replace JsonNode with shrinkable json nodes

+++ src/main/scala/de/leanovate/swaggercheck/formats/IntegerFormats.scala
@@ -5,20 +5,20 @@ import org.scalacheck.Gen
 
 object IntegerFormats {
 
-  object Int32 extends Format[BigDecimal] {
-    override def generate: Gen[BigDecimal] = Gen.choose(Int.MinValue, Int.MaxValue).map(BigDecimal.apply)
+  object Int32 extends Format[BigInt] {
+    override def generate: Gen[BigInt] = Gen.choose(Int.MinValue, Int.MaxValue).map(BigInt.apply)
 
-    override def verify(path: String, value: BigDecimal): VerifyResult =
+    override def verify(path: String, value: BigInt): VerifyResult =
       if (value.isValidInt)
         VerifyResult.success
       else
         VerifyResult.error(s"$value is not an int32: $path")
   }
 
-  object Int64 extends Format[BigDecimal] {
-    override def generate: Gen[BigDecimal] = Gen.choose(Long.MinValue, Long.MaxValue).map(BigDecimal.apply)
+  object Int64 extends Format[BigInt] {
+    override def generate: Gen[BigInt] = Gen.choose(Long.MinValue, Long.MaxValue).map(BigInt.apply)
 
-    override def verify(path: String, value: BigDecimal): VerifyResult =
+    override def verify(path: String, value: BigInt): VerifyResult =
       if (value.isValidLong)
         VerifyResult.success
       else
